refactor(quiz): reorganize package structure for grading and session management
