+++ src/main/java/com/vinaacademy/platform/feature/quiz/entity/Answer.java
@@ -1,12 +1,10 @@
 package com.vinaacademy.platform.feature.quiz.entity;
 
-import com.vinaacademy.platform.feature.common.entity.BaseEntity;
 import com.vinaacademy.platform.feature.common.entity.SoftDeleteEntity;
 import jakarta.persistence.*;
-import jakarta.persistence.Entity;
-import jakarta.persistence.Table;
 import lombok.*;
-import org.hibernate.annotations.*;
+import org.hibernate.annotations.SQLDelete;
+import org.hibernate.annotations.Where;
 
 import java.util.UUID;
