+++ src/main/java/com/vinaacademy/platform/feature/quiz/service/student/impl/QuizStudentServiceImpl.java
@@ -12,11 +12,11 @@ import com.vinaacademy.platform.feature.quiz.mapper.QuizMapper;
 import com.vinaacademy.platform.feature.quiz.mapper.QuizSubmissionMapper;
 import com.vinaacademy.platform.feature.quiz.repository.*;
 import com.vinaacademy.platform.feature.quiz.service.QuizCacheService;
+import com.vinaacademy.platform.feature.quiz.service.internal.QuizGradingService;
+import com.vinaacademy.platform.feature.quiz.service.internal.QuizSessionHandler;
+import com.vinaacademy.platform.feature.quiz.service.internal.QuizSubmissionFactory;
+import com.vinaacademy.platform.feature.quiz.service.internal.QuizValidator;
 import com.vinaacademy.platform.feature.quiz.service.student.QuizStudentService;
-import com.vinaacademy.platform.feature.quiz.service.student.internal.QuizGradingService;
-import com.vinaacademy.platform.feature.quiz.service.student.internal.QuizSessionHandler;
-import com.vinaacademy.platform.feature.quiz.service.student.internal.QuizSubmissionFactory;
-import com.vinaacademy.platform.feature.quiz.service.student.internal.QuizValidator;
 import com.vinaacademy.platform.feature.user.auth.annotation.RequiresResourcePermission;
 import com.vinaacademy.platform.feature.user.auth.helpers.SecurityHelper;
 import com.vinaacademy.platform.feature.user.constant.ResourceConstants;
