+++ src/main/java/com/vinaacademy/platform/feature/quiz/strategy/impl/TextGradingStrategy.java
@@ -1,10 +1,10 @@
-package com.vinaacademy.platform.feature.quiz.service.student.internal.strategy.impl;
+package com.vinaacademy.platform.feature.quiz.strategy.impl;
 
 import com.vinaacademy.platform.feature.quiz.dto.UserAnswerRequest;
 import com.vinaacademy.platform.feature.quiz.entity.Question;
 import com.vinaacademy.platform.feature.quiz.entity.UserAnswer;
 import com.vinaacademy.platform.feature.quiz.enums.QuestionType;
-import com.vinaacademy.platform.feature.quiz.service.student.internal.strategy.GradingStrategy;
+import com.vinaacademy.platform.feature.quiz.strategy.GradingStrategy;
 import org.springframework.stereotype.Component;
 
 @Component
